pyproject.toml
[project]
name = "gradient-skins"
version = "0.1.0"
description = "Minecraft gradient skin generator."
authors = [
    {name = "Xithrius", email = "xithrius@gmail.com"},
]
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pillow>=9.4.0",
    "numpy>=1.24.2"
]

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
lint = [
    "flake8>=6.0.0",
    "flake8-annotations>=3.0.0",
    "flake8-bugbear>=23.2.13",
    "flake8-docstrings>=1.7.0",
    "flake8-string-format>=0.3.0",
    "flake8-tidy-imports>=4.8.0",
    "flake8-todo>=0.7",
    "flake8-isort>=6.0.0",
    "pep8-naming>=0.13.3",
    "pyright>=1.1.293"
]
format = [
    "black>=23.1.0"
]
precommit = [
    "pre-commit>=3.0.4",
]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
gen = "python -m gradient-skins"
lint = "pre-commit run --all-files"
precommit = "pre-commit install"

[tool.black]
line-length = 80

[tool.pyright]
include = ["gradient-skins/**/*.py"]
exclude = ["**/__pycache__"]

typeCheckingMode = "strict"
# reportMissingTypeStubs = false
